+++ pkg/analysis_server/lib/src/services/correction/fix/dart/top_level_declarations.dart
@@ -39,7 +39,7 @@ class TopLevelDeclarations {
         continue;
       }
 
-      var libraryElement = analysisDriver.getLibraryByFile(file);
+      var libraryElement = await analysisDriver.getLibraryByFile(file);
       if (libraryElement == null) {
         continue;
       }