+++ pkg/analyzer_plugin/test/plugin/mocks.dart
@@ -159,30 +159,14 @@ class MockSource implements Source {
   @override
   TimestampedData<String> get contents => TimestampedData(0, '');
 
-  @Deprecated('Not used anymore')
-  @override
-  String get encoding => '';
-
   @override
   String get fullName => '/pkg/lib/test.dart';
-
-  @Deprecated('Use uri.isScheme("dart") instead')
-  @override
-  bool get isInSystemLibrary => false;
-
-  @Deprecated('Not used anymore')
-  @override
-  int get modificationStamp => 0;
 
   @override
   String get shortName => 'test.dart';
 
   @override
   Uri get uri => Uri.parse('package:test/test.dart');
-
-  @Deprecated('Use Source.uri instead')
-  @override
-  UriKind get uriKind => UriKind.PACKAGE_URI;
 
   @override
   bool exists() => true;